ints changes for size_t

+++ GA.cpp
@@ -1,17 +1,11 @@
-//
-// Created by Patrick Mintram on 22/07/2017.
-//
-
 #include <unordered_map>
 #include "GA.h"
-#include <vector>
-
 
 tour_t GA::tournament_selection( population_t &population )
 {
 	population_t pop(_tournament_size, false);
-	for(int i = 0; i < _tournament_size; i++){
-		int random = std::rand() % population.get_size();
+	for(size_t i = 0; i < _tournament_size; i++){
+		size_t random = std::rand() % population.get_size();
 		pop.set_tour(i, population.get_tour(random));
 	}
 
@@ -21,18 +15,18 @@ tour_t GA::tournament_selection( population_t &population )
 
 tour_t &GA::crossover( tour_t &p1, tour_t &p2, tour_t &child)
 {
-	int start = int(rand() % p1.get_city_count());
-	int end = int((rand() % (p1.get_city_count() - start)) + start);
+	size_t start = (size_t)(rand() % p1.get_city_count());
+	size_t end = (rand() % (p1.get_city_count() - start)) + start;
 
 	//std::cout << "xover chosen " << start << " to " << end << std::endl;
 
-	for(int i = start; i < end; i++){
+	for(size_t i = start; i < end; i++){
 		child.set_city(i, p1.get_city(i));
 	}
 
-	for( int i = 0; i < p2.get_city_count(); i++){
+	for( size_t i = 0; i < p2.get_city_count(); i++){
 		if(!child.contains_city(p2.get_city(i))){
-			for(int ii = 0; ii < child.get_city_count(); ii++){
+			for(size_t ii = 0; ii < child.get_city_count(); ii++){
 				if(child.get_city(ii).get_name() == empty_city_t::EMPTY){
 					child.set_city(ii, p2.get_city(i));
 					break;
@@ -46,13 +40,13 @@ tour_t &GA::crossover( tour_t &p1, tour_t &p2, tour_t &child)
 
 population_t GA::evolve_population(population_t& input_pop){
 	population_t next_pop(input_pop.get_size(), false);
-	int elitism_offset = 0;
+	size_t elitism_offset = 0;
 	if(_elitist){
 		elitism_offset++;
 		next_pop.set_tour(0, input_pop.get_fittest());
 	}
 
-	for(int i = elitism_offset; i < input_pop.get_size(); i++){
+	for(size_t i = elitism_offset; i < input_pop.get_size(); i++){
 		tour_t child;
 		tour_t p1 = tournament_selection(input_pop);
 		tour_t p2 = tournament_selection(input_pop);
@@ -75,10 +69,10 @@ population_t GA::evolve_population(population_t& input_pop){
 
 void GA::mutate( tour_t &t )
 {
-	for(int i = 0; i < t.get_city_count(); i++){
+	for(size_t i = 0; i < t.get_city_count(); i++){
 		int random = std::rand();
 		if(random < _mutation_rate){
-			int rand_idx = int(std::rand() % t.get_city_count());
+			size_t rand_idx = (size_t)(std::rand() % t.get_city_count());
 			city_t c1 = t.get_city(i);
 			city_t c2 = t.get_city(rand_idx);
 
@@ -119,13 +113,14 @@ std::string GA::to_string(void) const {
 double GA::get_mutation_rate(void) {
 	return _mutation_rate;
 }
-int GA::get_tournament_size(void) {
+size_t GA::get_tournament_size( void ) {
 	return _tournament_size;
 }
 bool GA::is_elitist(void){
 	return _elitist;
 }
-void GA::set_tournament_size(const int s) {
+
+void GA::set_tournament_size(const size_t s) {
 	_tournament_size = s;
 }
 void GA::set_mutation_rate(const double m) {
